+++ ios/MullvadVPN/View controllers/Tunnel/ConnectionView/ChipView/ChipView.swift
@@ -45,6 +45,8 @@ struct ChipView: View {
                 )
                 .padding(borderWidth)
         )
+        .accessibilityIdentifier(item.id.accessibilityId)
+        .accessibilityLabel(item.name)
     }
 }